Remove local storage methods. Update click event in carousel to update windlow location hash. Update componentDidMount in recipe to grab recipe id from window hash and send a fetch request to spoonacular api. #3
